requirements.txt
hydra-core==1.3.2
matplotlib==3.7.2
numpy==2.2.2
omegaconf==2.3.0
Pillow==11.1.0
pydantic==2.10.5
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.15.1
sionna==0.18.0
tensorflow==2.15.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
wandb==0.15.11
